refactor: major components folder cleanup and project structure alignment
