+++ src/components/Homepage/Sections/HomepageContent.js
@@ -1,12 +1,12 @@
-import FightGuideList from "./FightGuideList";
-import Offerings from "./OfferingCards";
+import FightGuideList from "../Widgets/FightGuideList";
+import Offerings from "../Widgets/OfferingCards";
 
 export default function HomepageContent() {
   return (
-    <div className="max-w-screen-2xl mx-auto p-10 grid gap-y-10">
+    <div className="grid p-10 mx-auto max-w-screen-2xl gap-y-10">
       <FightGuideList />
       <Offerings />
-      <p className="text-center text-xl">
+      <p className="text-xl text-center">
         If you are interested in high-end content on NA Data Centers and what we
         have to offer, come be a part of our community!
       </p>
