+++ src/components/Mdx/Elements/Details.js
