+++ src/components/Mdx/Layout/MdxLayout.js
@@ -1,9 +1,6 @@
 import "@/app/globals.css";
-import { Box } from "@mui/material";
-import NavBar from "../NavBar/NavBar";
-import Footer from "../Footer/Footer";
+
 import { Roboto } from "next/font/google";
-import MUITheme from "@/app/theme";
 
 // note: this can be removed when MDX rendering is moved to @/app/ (NAUR-84)
 const roboto = Roboto({
