+++ src/components/Mdx/Utils/CopyToClipboard.js
@@ -1,6 +1,7 @@
 "use client";
-import React from "react";
+
 import { ContentCopy } from "@mui/icons-material";
+import React from "react";
 
 export const CopyToClipboard = ({ children }) => {
   const textInput = React.useRef(null);
@@ -15,7 +16,7 @@ export const CopyToClipboard = ({ children }) => {
       <button
         aria-label="Copy code"
         type="button"
-        className="absolute right-2 top-2 w-10 h-10 p-1 rounded hover:bg-gray-700 active:bg-gray-600 transition-colors"
+        className="absolute w-10 h-10 p-1 transition-colors rounded right-2 top-2 hover:bg-gray-700 active:bg-gray-600"
         onClick={onCopy}
       >
         <ContentCopy />
