+++ src/components/Mdx/Utils/TableOfContents.js
@@ -1,5 +1,6 @@
-import ScrollSpy from "react-scrollspy-navigation";
 import ArrowDropDownIcon from "@mui/icons-material/ArrowDropDown";
+import ScrollSpy from "react-scrollspy-navigation";
+
 // recursively crawls through the tree generated by @stefanprobst/rehype-extract-toc
 // and generates the table of contents
 function recursiveToc(toc, collapse, level = 0) {
