+++ src/lib/assets.js
@@ -1,12 +1,21 @@
 // icons
+import Calendar from "@/assets/Icons/Calendar.png";
 import Chat from "@/assets/Icons/Chat.png";
-import Lightbulb from "@/assets/Icons/Lightbulb.png";
 import Contact from "@/assets/Icons/Contact.png";
 import Discord from "@/assets/Icons/Discord.png";
-import Calendar from "@/assets/Icons/Calendar.png";
 import Github from "@/assets/Icons/githublogo.png";
+import Lightbulb from "@/assets/Icons/Lightbulb.png";
 import Naur from "@/assets/Icons/naur_icon.png";
+import Pandora from "@/assets/Images/Pandora.avif";
 import Patreon from "@/assets/Icons/Patreon.png";
+// MUI icons
+import PersonSearchIcon from "@mui/icons-material/PersonSearch";
+// homepage card images
+import m5s from "@/assets/Images/m5s.avif";
+import m6s from "@/assets/Images/m6s.avif";
+import m7s from "@/assets/Images/m7s.avif";
+import m8s from "@/assets/Images/m8s.avif";
+import recollection from "@/assets/Images/recollection.avif";
 export const icons = {
   Chat,
   Lightbulb,
@@ -18,16 +27,6 @@ export const icons = {
   Patreon,
 };
 
-// homepage card images
-import Pandora from "@/assets/Images/Pandora.avif";
-import m5s from "@/assets/Images/m5s.avif";
-import m6s from "@/assets/Images/m6s.avif";
-import m7s from "@/assets/Images/m7s.avif";
-import m8s from "@/assets/Images/m8s.avif";
-import recollection from "@/assets/Images/recollection.avif";
-
 export const images = { Pandora, m5s, m6s, m7s, m8s, recollection };
 
-// MUI icons
-import PersonSearchIcon from "@mui/icons-material/PersonSearch";
 export const iconsMui = { PersonSearchIcon };
